% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_equilibrium.R
\name{find_equilibrium}
\alias{find_equilibrium}
\title{Find the demographic equilibrium}
\usage{
find_equilibrium(xres, pars, init)
}
\arguments{
\item{xres}{The resident trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of two starting values for solving of the demographic
equilibrium}
}
\value{
A vector of two equilibrium population sizes, one for each habitat.
}
\description{
Solves the demographic equilibrium equation for the population sizes in the
two habitats
}
\details{
This function uses a multivariate numerical root finding algorithm
(\code{?pracma::fsolve}) to find solutions to the equation.
}
\examples{

pars <- get_default_pars()
find_equilibrium(0, pars, init = rep(1000, 2))

}
