% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_equilibrium_burnin.R
\name{find_equilibrium_burnin}
\alias{find_equilibrium_burnin}
\title{Find the burn-in demographic equilibrium}
\usage{
find_equilibrium_burnin(xres, pars)
}
\arguments{
\item{xres}{The resident trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}
}
\value{
The equilibrium population size
}
\description{
Solves the burn-in demographic equilibrium equation for the population size,
given parameter values.
}
\details{
This function uses an analytical expression for the non-extinct
solution to the equation (the other solution being when population size
is zero).
}
\examples{

pars <- get_default_pars()
find_equilibrium_burnin(0, pars)

}
