% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_equilibrium_di.R
\name{find_equilibrium_di}
\alias{find_equilibrium_di}
\title{Find the dimorphic demographic equilibrium}
\usage{
find_equilibrium_di(xres1, xres2, pars, init)
}
\arguments{
\item{xres1}{The resident trait value in the first morph}

\item{xres2}{The resident trait value in the second morph}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of four starting values for solving of the demographic
equilibrium}
}
\value{
A vector of four equilibrium population sizes: morph 1 in habitat 1,
morph 1 in habitat 2, morph 2 in habitat 1 and morph 2 in habitat 2.
}
\description{
Solves the demographic equilibrium equation for the population sizes in the
two morphs and the two habitats, after branching has occurred, given
parameter values.
}
\details{
This function uses a multivariate numerical root finding algorithm
(\code{?pracma::fsolve}) to find solutions to the equation.
}
\examples{

pars <- get_default_pars()
find_equilibrium_di(-0.001, 0.001, pars, init = rep(1000, 4))

}
