% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_roots.R
\name{find_roots}
\alias{find_roots}
\title{Numerically find roots of a function}
\usage{
find_roots(f, from, to, n = 10, size = 1, precis = NULL, ...)
}
\arguments{
\item{f}{The function whose roots to find (must have one argument only)}

\item{from, to}{The bounds of the range of values to explore}

\item{n}{The number of bins in which to split the range}

\item{size}{The size of the sliding window, in number of bins}

\item{precis}{To what precision digit to round the solutions (no rounding if
\code{NULL})}

\item{...}{Additional arguments to be passed to \code{uniroot_noerr}}
}
\value{
A vector of solutions
}
\description{
Root-finding algorithm using sliding windows to numerically find the roots
of a function within a range.
}
\details{
The function numerically looks for one root within each window, then
removes duplicate solutions and returns the output. Sometimes duplicate
solutions may not be recognized as such because they differ by a negligible
amount. Use the \code{precis} argument to round those solutions to a lower
precision (done using \code{round}) and recognize them as duplicates.
}
\examples{

f <- function(x) x^2 - 1
find_roots(f, -10, 10)

}
\seealso{
\code{get_sliding_windows}, \code{uniroot_noerr}
}
