% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Predict evolutionary endpoint}
\usage{
get_endpoint(
  xstart,
  pars,
  init,
  from = xstart - 10,
  to = xstart + 10,
  step = 1e-04,
  ...
)
}
\arguments{
\item{xstart}{The starting resident trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of two starting values for solving of the demographic
equilibrium}

\item{from, to}{Arguments to be passed to \code{find_singularities} (only
the bound towards which the initial selection gradient is pointing is
considered)}

\item{step}{Argument to be passed to \code{is_convergent}}

\item{...}{Optional extra arguments for \code{find_singularities}}
}
\value{
A data frame with the singularity, whether it is attainable and
whether it is evolutionarily stable. If a singularity is attainable and
stable, it is an evolutionarily stable strategy (labelled "ESS"). If it is
attainable and unstable, it is a branching point ("BP"). If it is
unattainable, it is a repeller.
}
\description{
Predicts the evolutionary equilibrium reached given a starting point and
some parameter values, and if this equilibrium is stable.
}
\details{
The function computes the selection gradient at the starting point
to know in which direction evolution is leading, then finds the closest
singularity in that direction.
}
\examples{

pars <- get_default_pars()
get_endpoint(0, pars, init = rep(1000, 2))

}
\seealso{
\code{get_gradient}, \code{find_singularities}, \code{is_stable},
\code{is_convergent}
}
