% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint_within.R
\name{get_endpoint_within}
\alias{get_endpoint_within}
\title{Predict within-habitat evolutionary endpoint}
\usage{
get_endpoint_within(
  xstart,
  pars,
  from = xstart - 10,
  to = xstart + 10,
  step = 1e-04,
  ...
)
}
\arguments{
\item{xstart}{The starting resident trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{from, to}{Arguments to be passed to \code{find_singularities_within}
(only the bound towards which the initial selection gradient is pointing is
considered)}

\item{step}{Argument to be passed to \code{is_convergent_within}}

\item{...}{Optional extra arguments for \code{find_singularities_within}}
}
\value{
A data frame with the singularity, whether it is attainable and
whether it is evolutionarily stable. If a singularity is attainable and
stable, it is an evolutionarily stable strategy (labelled "ESS"). If it is
attainable and unstable, it is a branching point ("BP"). If it is
unattainable, it is a repeller.
}
\description{
Predicts the evolutionary equilibrium reached given a starting point and
some parameter values, and if this equilibrium is stable.
}
\details{
The function computes the selection gradient at the starting point
to know in which direction evolution is leading, then finds the closest
singularity in that direction.
}
\examples{

pars <- get_default_pars()
get_endpoint_within(0, pars)

}
\seealso{
\code{get_gradient_within}, \code{find_singularities_within},
\code{is_stable_within}, \code{is_convergent_within}
}
