% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_pars.R
\name{get_example_pars}
\alias{get_example_pars}
\title{Example parameter set}
\usage{
get_example_pars()
}
\value{
A list of expressions
}
\description{
Generates an unevaluated list of parameter setting statements
}
\note{
Use \code{eval()} on an expression in the list to execute it.
Beware that evaluation will run each statement, thereby overriding any
variable that shares its name with one of the parameters.
}
\examples{

get_example_pars()

}
\seealso{
\code{get_default_pars}
}
