% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gradient_burnin.R
\name{get_gradient_burnin}
\alias{get_gradient_burnin}
\title{Burn-in selection gradient}
\usage{
get_gradient_burnin(xres, pars)
}
\arguments{
\item{xres}{The resident trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}
}
\value{
The selection gradient
}
\description{
Computes the selection gradient during the burn-in period given parameter
values.
}
\details{
The functions uses \code{find_equilibrium_burnin} to compute the
equilibrium population size given the resident trait value, which is needed
to compute the gradient.
}
\examples{

pars <- get_default_pars()
get_gradient_burnin(0, pars)

}
\seealso{
\code{find_equilibrium_burnin}
}
