% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gradient_di.R
\name{get_gradient_di}
\alias{get_gradient_di}
\title{Dimorphic selection gradients}
\usage{
get_gradient_di(xres1, xres2, pars, init)
}
\arguments{
\item{xres1}{The resident trait value in the first morph}

\item{xres2}{The resident trait value in the second morph}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of four starting values for solving of the demographic
equilibrium}
}
\value{
A vector of two selection gradients, one for each morph
}
\description{
Computes the selection gradients of the two morphs after branching has
occurred.
}
\details{
The functions uses \code{find_equilibrium_di} to compute the
equilibrium population sizes (in the two habitats and for the two morphs)
given the resident trait values, which are needed to compute the gradients.
}
\examples{

pars <- get_default_pars()
get_gradient_di(-0.001, 0.001, pars, init = rep(1000, 4))

}
\seealso{
\code{find_equilibrium_di}
}
