% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gradient_within.R
\name{get_gradient_within}
\alias{get_gradient_within}
\title{Within-habitat selection gradient}
\usage{
get_gradient_within(xres, pars)
}
\arguments{
\item{xres}{The resident trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}
}
\value{
The selection gradient
}
\description{
Computes the selection gradient given a resident value and model parameters.
}
\details{
The functions uses \code{find_equilibrium_within} to compute the
equilibrium population size given the resident trait value
}
\examples{

pars <- get_default_pars()
get_gradient_within(0, pars)

}
\seealso{
\code{find_equilibrium_within}, \code{get_gradient}
}
