% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lambda.R
\name{get_lambda}
\alias{get_lambda}
\title{Invasion fitness}
\usage{
get_lambda(x, xres, pars, init, fast = FALSE)
}
\arguments{
\item{x}{The mutant trait value}

\item{xres}{The resident trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of two starting values for solving of the demographic
equilibrium}
}
\value{
The invasion fitness of the mutant given the resident
}
\description{
Computes the invasion fitness of a mutant in a resident population, given
trait values and model parameters.
}
\details{
This function computes an expression for the leading eigenvalue of
the transition matrix of the model, which is the invasion fitness of the
mutant.

The function will find the demographic equilibrium of the resident (passing
\code{init} to \code{find_equilibrium}) only if \code{fast = FALSE}.
Otherwise, \code{init} is interpreted as a vector of equilibrium population
sizes (useful when computing the invasion fitness of many mutants in the
same resident population, e.g. in \code{plot_pip}).
}
\examples{

pars <- get_default_pars()
get_lambda(0.01, 0, pars, init = rep(1000, 2))

}
\seealso{
\code{plot_pip}, \code{find_equilibrium}
}
