% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Model set-up}
\usage{
get_model()
}
\value{
A list of expressions to run to evaluate the model
}
\description{
Generates an unevaluated list of model variable statements
}
\note{
Use \code{eval()} on an expression in the list to execute it.
Beware that the evaluation of the model assumes that parameters and their
values are present in the environment (see e.g. \code{?get_default_pars}) and
that variables \code{x} and \code{xres} exist.
}
\examples{

get_model()

}
\seealso{
\code{get_model_di}, \code{get_model_burnin}
}
