% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_within.R
\name{get_model_within}
\alias{get_model_within}
\title{Within-habitat model set-up}
\usage{
get_model_within()
}
\value{
A list of expressions to run to evaluate the model
}
\description{
Generates an unevaluated list of model variable statements
}
\note{
Use \code{eval()} on an expression in the list to execute it.
Beware that the evaluation of the model assumes that parameters and their
values are present in the environment (see e.g. \code{?get_default_pars}) and
that variables \code{x} and \code{xres} exist.
}
\examples{

get_model_within()

}
\seealso{
\code{get_model}, \code{get_model_di}
}
