% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stable.R
\name{is_stable}
\alias{is_stable}
\title{Evaluate evolutionary stability}
\usage{
is_stable(xeq, pars, init)
}
\arguments{
\item{xeq}{Some equilibrium trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of two starting values for solving of the demographic
equilibrium}
}
\value{
A boolean
}
\description{
Assesses whether an equilibrium trait value is evolutionarily stable, or
invasion-proof.
}
\details{
This functions computes an expression for the second derivative
of the invasion fitness function with respect to the mutant trait value,
evaluated at the resident trait value at equilibrium \code{xeq}. If this
value is negative then the equilibrium is a fitness maximum, i.e. it is
evolutionarily stable. It is unstable otherwise.
}
\examples{

pars <- get_default_pars()
is_stable(0, pars, init = rep(1000, 2))

}
\seealso{
\code{is_convergent}
}
