% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stable_within.R
\name{is_stable_within}
\alias{is_stable_within}
\title{Evaluate within-habitat evolutionary stability}
\usage{
is_stable_within(xeq, pars, value = FALSE)
}
\arguments{
\item{xeq}{Some equilibrium trait value}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{value}{Whether to return the actual value of the curvature of the
fitness function (defaults to \code{FALSE}, i.e. only returns whether this
value is negative)}
}
\value{
A boolean
}
\description{
Assesses whether an equilibrium trait value is evolutionarily stable, or
invasion-proof.
}
\details{
This functions computes an expression for the second derivative
of the invasion fitness function with respect to the mutant trait value,
evaluated at the resident trait value at equilibrium \code{xeq}. If this
value is negative then the equilibrium is a fitness maximum, i.e. it is
evolutionarily stable. It is unstable otherwise.
}
\examples{

pars <- get_default_pars()
is_stable_within(0, pars)

}
\seealso{
\code{is_convergent_within}, \code{is_stable}
}
