% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pip.R
\name{plot_pip}
\alias{plot_pip}
\title{Pairwise invasibility plot}
\usage{
plot_pip(xvalues, pars, init, binary = TRUE, plotit = TRUE)
}
\arguments{
\item{xvalues}{A vector of trait values over which to compute the invasion
fitness. Each value will be taken both as a mutant and as a resident trait
value, so all pairwise combinations will be evaluated.}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of two starting values for solving of the demographic
equilibrium}

\item{binary}{Whether or not to plot whether the mutant invades (\code{TRUE})
or the actual value of the invasion fitness (\code{FALSE})}

\item{plotit}{Whether to return a plot or the underlying tibble}
}
\value{
A \code{ggplot} object
}
\description{
Generates a PIP showing the combinations of mutant and resident trait values
for which the mutant invades the resident (i.e. has an invasion fitness
greater than one).
}
\details{
The function computes the invasion fitness for all pairwise
combinations of mutants and residents, but recycles equilibrium population
sizes computed for each resident to gain in speed (using \code{fast = TRUE}
in \code{get_lambda}).
}
\examples{

pars <- get_default_pars()
plot_pip(seq(-1, 1, 1), pars, init = rep(1000, 2))

}
\seealso{
\code{get_lambda}, \code{find_equilibrium}
}
