% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_burnin.R
\name{simulate_burnin}
\alias{simulate_burnin}
\title{Simulate burn-in}
\usage{
simulate_burnin(xstart, ntimes, pars, mu = 0.01, sigma = 0.1)
}
\arguments{
\item{xstart}{Starting trait value}

\item{ntimes}{The number of time steps to simulate for}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{mu}{The mutation rate of the trait (per time step)}

\item{sigma}{The (standard deviation of) mutational step size upon mutation}
}
\value{
A tibble containing the trait value of the population at each time
point
}
\description{
Simulates trait evolution during the burn-in period.
}
\details{
See \code{?simulate}.
}
\examples{

pars <- get_default_pars()
simulate_burnin(0, 10, pars)

}
\seealso{
\code{find_equilibrium_burnin}, \code{get_gradient_burnin}
}
