% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_di.R
\name{simulate_di}
\alias{simulate_di}
\title{Simulate dimorphic evolution}
\usage{
simulate_di(xstart, ntimes, pars, init, mu = 0.01, sigma = 0.1, dodge = 0.001)
}
\arguments{
\item{xstart}{Starting trait value}

\item{ntimes}{The number of time steps to simulate for}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of four starting values for solving of the demographic
equilibrium}

\item{mu}{The mutation rate of the trait (per time step)}

\item{sigma}{The (standard deviation of) mutational step size upon mutation}

\item{dodge}{How much to displace each morph from the branching point once
branching has occurred}
}
\value{
A tibble containing the trait values in the two morphs at each time
point.
}
\description{
Simulates trait evolution after branching.
}
\details{
See \code{?simulate}.
}
\examples{

pars <- get_default_pars()
simulate_di(0, 10, pars, init = rep(1000, 4))

}
\seealso{
\code{find_equilibrium_di}, \code{get_gradient_di}
}
