% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot_noerr.R
\name{uniroot_noerr}
\alias{uniroot_noerr}
\title{Root finding without error}
\usage{
uniroot_noerr(f, interval, ...)
}
\arguments{
\item{f}{The function whose roots to find (must have one argument only)}

\item{interval}{The range within which to search for a root}

\item{...}{Additional parameters for \code{uniroot}}
}
\value{
A root, or \code{NULL}
}
\description{
Runs the \code{uniroot} root finding algorithm and returns \code{NULL}
instead of erroring if no solution is found.
}
\details{
The \code{uniroot} function errors if no solution is found, which
is undesirable when looking for multiple roots across a range, or when
looking for roots as part of a bigger pipeline. This function catches the
error produced by \code{uniroot} if no solution is found and returns
\code{NULL} instead.
}
\examples{

f <- function(x) x^2 - 1
uniroot_noerr(f, c(-50, -30))

}
\seealso{
\code{uniroot}
}
