% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mono.R
\name{simulate_mono}
\alias{simulate_mono}
\title{Simulate monomorphic evolution}
\usage{
simulate_mono(
  xstart,
  ntimes,
  pars,
  init,
  mu = 0.01,
  sigma = 0.1,
  branch = TRUE,
  tol = 1e-04
)
}
\arguments{
\item{xstart}{Starting trait value}

\item{ntimes}{The number of time steps to simulate for}

\item{pars}{An unevaluated parameter-list (e.g. as returned by
\code{get_default_pars})}

\item{init}{A vector of two starting values for solving of the demographic
equilibrium}

\item{mu}{The mutation rate of the trait (per time step)}

\item{sigma}{The (standard deviation of) mutational step size upon mutation}

\item{branch}{Whether to end the simulation when a branching point has
been reached}

\item{tol}{The tolerance when assessing if the selection gradient is close
enough to zero and an equilibrium has been reached}
}
\value{
A tibble containing the trait value of the population at each time
point
}
\description{
Simulates trait evolution in a monomorphic population.
}
\details{
See \code{?simulate}.
}
\examples{

pars <- get_default_pars()
simulate_mono(-1, 10, pars, init = rep(1000, 2))

}
\seealso{
\code{find_equilibrium}, \code{get_gradient}
}
